pyproject.toml
[project]
name = "nad"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "contextily>=1.6.2",
    "ipykernel>=6.30.0",
    "matplotlib>=3.10.5",
    "numpy>=1.23.4,<2.0.0",
    "openpyxl>=3.1.5",
    "osmnx>=2.0.5",
    "pandana>=0.7",
    "pandas-stubs>=2.3.0.250703",
    "pykrige>=1.7.2",
    "pca>=2.10.0",
    "scipy>=1.16.1",
    "mapclassify>=2.10.0",
    "xlrd>=2.0.2",
    "photutils>=2.2.0",
    "rasterio>=1.4.3",
    "astropy>=7.1.0",
    "binsreg>=2.1.7",
]

[tool.ruff.lint]
select = ["ALL"]
ignore = ["D", "PLR2004"]

[tool.basedpyright]
typeCheckingMode = "standard"
ignore = ["**/*.ipynb"]
